Silence MSVC compiler warning
include\generate_phi.hpp(138): warning C4244: 'argument': conversion
from 'int64_t' to 'int32_t', possible loss of data

+++ include/imath.hpp
@@ -21,9 +21,9 @@
 
 namespace {
 
-inline int64_t isquare(int32_t x)
+inline int64_t isquare(int64_t x)
 {
-  return x * (int64_t) x;
+  return x * x;
 }
 
 template <typename A, typename B>
